:sparkles: Add draw_ray with angle
